Moved ImagePreview files to ProjectPhotos folder
